backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Transcription
openai-whisper==20231117

# YouTube download
yt-dlp

# Audio processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
